better data fetching and selecting
